+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
+- Drop the CSV export of events (#3606)
 - Drop obsolete entries from `ext_emconf.php` (#3597)
 - Remove the salutation mode setting from the flexforms (#3591)
 - Drop the autologout for no-longer-supported onetimeaccount logins (#3556)